+++ app/smartbrain_3000/web/_app/immutable/chunks/CKEjtgXk.js
@@ -1,0 +1,1 @@
+import"./CB3BI3v6.js";